requirements.txt
python-dotenv==0.20.0
requests==2.31.0
httpx==0.26.0